refactor: move delete parseing work out of influxdb_line_protocol crate
