+++ influxdb_line_protocol/src/lib.rs
@@ -16,8 +16,6 @@
     clippy::clone_on_ref_ptr
 )]
 
-pub mod delete_parser;
-
 use fmt::Display;
 use nom::{
     branch::alt,
