+++ query/Cargo.toml
@@ -24,16 +24,20 @@ datafusion = { path = "../datafusion" }
 datafusion_util = { path = "../datafusion_util" }
 futures = "0.3"
 hashbrown = "0.11"
+influxdb_line_protocol = { path = "../influxdb_line_protocol" }
 internal_types = { path = "../internal_types" }
 metrics = { path = "../metrics" }
+observability_deps = { path = "../observability_deps" }
 parking_lot = "0.11.2"
 regex = "1"
 snafu = "0.6.3"
+sqlparser = "0.10.0"
+thiserror = "1.0.28"
 tokio = { version = "1.11", features = ["macros"] }
 tokio-stream = "0.1.2"
 trace = { path = "../trace" }
 tracker = { path = "../tracker" }
-observability_deps = { path = "../observability_deps" }
+
 
 # use libc on unix like platforms to set worker priority in DedicatedExecutor
 [target."cfg(unix)".dependencies.libc]
